requirements12.txt
asgiref==3.2.5
Django==3.0.4
pkg-resources==0.0.0
python-dotenv==0.12.0
pytz==2019.3
random-name==0.1.1
sqlparse==0.3.1
